+++ taskflowbackend/flows/landing_zone_create.py
@@ -19,8 +19,8 @@ class Flow(BaseLinearFlow):
         return super().validate()
 
     def build(self, force_fail=False):
-        project_group = self.irods_backend.get_user_group_name(self.project)
-        owner_group = self.irods_backend.get_user_group_name(self.project, True)
+        project_group = self.irods_backend.get_group_name(self.project)
+        owner_group = self.irods_backend.get_group_name(self.project, True)
         zone_root = self.irods_backend.get_zone_path(self.project)
         zone = LandingZone.objects.get(sodar_uuid=self.flow_data['zone_uuid'])
         user_path = os.path.join(zone_root, zone.user.username)