+++ taskflowbackend/flows/role_delete.py
@@ -11,7 +11,7 @@ class Flow(BaseLinearFlow):
         return super().validate()
 
     def build(self, force_fail=False):
-        existing_group = self.irods_backend.get_user_group_name(self.project)
+        existing_group = self.irods_backend.get_group_name(self.project)
 
         self.add_task(
             irods_tasks.RemoveUserFromGroupTask(