+++ CHANGELOG.rst
@@ -18,7 +18,7 @@ Added
 - **Irodsbackend**
     - ``get_objects()`` checksum support (#2038)
     - ``get_objects()`` offset support (#1997)
-    - ``get_user_group_name()`` owner/delegate group support (#2109)
+    - ``get_group_name()`` owner/delegate group support (#2109)
 - **Landingzones**
     - Site read-only mode support (#2051)
     - File type prohibiting by file name suffix (#2064)
@@ -51,6 +51,7 @@ Changed
     - Upgrade to gunicorn v23 (#2068)
 - **Irodsbackend**
     - Allow use of ``include_md5`` and ``limit`` together in ``get_objs_recursively()`` (#1887)
+    - Rename ``get_user_group_name()`` to ``get_group_name()`` (#2121)
 - **Landingzones**
     - Define app settings as ``PluginAppSettingDef`` objects (#2051)
     - Do not mute zone title and description with busy zones (#2092)